added check station functionality
introduced the check station functionality that checks the equipment added to the scene against a preset of stations. The preset station was selected in the beginning of the application. Added another singleton that carries forward with do not destroy on load to make sure the user selected station and the user setup station are same and provide feedback to user